:boom: eww(bar/workspaces): fix workspace indicator with a literal script
