+++ eww/eww.yuck
@@ -1,20 +1,10 @@
+; Variables
+(include "variables.yuck")
+
 ; Windows
-(include "./windows/calendar.yuck")
-(include "./windows/control-center.yuck")
-(include "./windows/bar.yuck")
-(include "./windows/powermenu.yuck")
-(include "./windows/audio-popup.yuck")
-
-; Bar widgets
-(include "./widgets/bar/workspaces.yuck")
-(include "./widgets/bar/clock.yuck")
-(include "./widgets/bar/control-center.yuck")
-(include "./widgets/bar/audio.yuck")
-(include "./widgets/bar/media.yuck")
-(include "./widgets/bar/logo.yuck")
-(include "./widgets/bar/window.yuck")
-(include "./widgets/bar/network.yuck")
-(include "./widgets/bar/battery.yuck")
-
-; Control Center widgets
-
+(include "windows/calendar.yuck")
+(include "windows/control-center.yuck")
+(include "windows/bar.yuck")
+(include "windows/powermenu.yuck")
+(include "windows/audio-popup.yuck")
+(include "windows/notification-popup.yuck")
