--- eww/scripts/notification-history.sh
@@ -1,10 +1,0 @@
-#!/usr/bin/env bash
-
-prev_history=$(dunstctl history)
-
-while true; do
-    if ! [[ $prev_history == $(dunstctl history) ]]; then
-        prev_history=$(dunstctl history)
-        echo "$(echo $prev_history | jq -c '.data.[]')"
-    fi
-done
