+++ eww/widgets/audio-popup/source-slider.yuck
@@ -5,7 +5,7 @@
         (overlay
             (scale :min 0
                    :max 100
-                   :value "${volume_json.source}"
+                   :value "${json_volume.source}"
                    :orientation "horizontal"
                    :draw-value false
                    :flipped false
