+++ eww/widgets/bar/audio.yuck
@@ -1,15 +1,12 @@
 
-(deflisten json_audio :initial `{ "output": 35, "source": 80 }`
-`sh ./scripts/get-volume-watch.sh`)
-
 (defwidget audio []
     (eventbox :onclick "eww open --toggle audio-popup"
               :class "audio-eventbox"
         (box :class "audio"
             (eventbox :onscroll `[ {} == "up" ] && wpctl set-volume @DEFAULT_AUDIO_SINK@ 5%+ || wpctl set-volume @DEFAULT_AUDIO_SINK@ 5%-`
-                (label :text "${ json_audio.output != 0 ? '󰕾' : '󰝟' }  ${json_audio.output}%"))
+                (label :text "${ json_volume.output != 0 ? '󰕾' : '󰝟' }  ${json_volume.output}%"))
             (eventbox :onscroll `[ {} == "up" ] && wpctl set-volume @DEFAULT_AUDIO_SOURCE@ 5%+ || wpctl set-volume @DEFAULT_AUDIO_SOURCE@ 5%-`
-                (label :text "${ json_audio.source != 0 ? '󰍬' : '󰍭' }  ${json_audio.source}%"))
+                (label :text "${ json_volume.source != 0 ? '󰍬' : '󰍭' }  ${json_volume.source}%"))
         )
     )
 )
