+++ eww/widgets/control-center/quickactions.yuck
@@ -1,4 +1,5 @@
-(deflisten hostname :initial "GNU/Linux"
+(defpoll hostname :initial "GNU/Linux"
+                  :interval "24h"
 `cat /etc/hostname`)
 
 (defpoll uptime_info :interval "50s"
@@ -13,7 +14,7 @@
              :halign "start"
 
             (label :xalign 0
-                   :text "  ${hostname}"
+                   :text "   ${hostname}"
                    :class "hostname")
 
             (label :xalign 0
@@ -29,6 +30,9 @@
              (button :class "lock"
                      :onclick "hyprctl dispatch exec hyprlock"
              "󰌾")
+             (button :class "color-picker"
+                     :onclick "sh $HOME/.config/eww/scripts/color-picker.sh"
+             "󰴱")
              (button :class "powermenu"
                      :onclick "eww close-all; eww open powermenu"
              "󰗽")
