+++ eww/windows/audio-popup.yuck
@@ -1,9 +1,6 @@
 (include "./widgets/audio-popup/output-slider.yuck")
 (include "./widgets/audio-popup/source-slider.yuck")
 
-(deflisten volume_json :initial `{ "output": 60, "source": 80 }`
-`sh ./scripts/get-volume-watch.sh`)
-
 (defwindow audio-popup []
       :monitor 0
       :namespace "eww-audio"
@@ -26,7 +23,7 @@
         (box :class "vertical-button-row"
             (button :class "more-settings"
                     :onclick "eww close audio-popup; hyprctl dispatch exec pavucontrol"
-            (label :text "Show on Volume Control"
+            (label :text "More devices"
                    :xalign 0))
         )
     )
