+++ eww/windows/bar.yuck
@@ -1,3 +1,13 @@
+(include "./widgets/bar/workspaces.yuck")
+(include "./widgets/bar/clock.yuck")
+(include "./widgets/bar/cc-toggle.yuck")
+(include "./widgets/bar/audio.yuck")
+(include "./widgets/bar/media.yuck")
+(include "./widgets/bar/logo.yuck")
+(include "./widgets/bar/window.yuck")
+(include "./widgets/bar/network.yuck")
+(include "./widgets/bar/battery.yuck")
+
 (defwindow bar 
     :monitor 0
     :geometry (geometry :width "100%"
@@ -33,7 +43,7 @@
         (audio)
         (battery)
         (network)
-        (control-center-toggle)
+        (cc-toggle)
       )
     )
 )
