--- hypr/autostart.conf
@@ -1,21 +1,0 @@
-
-###############
-## AUTOSTART ##
-###############
-
-exec-once = /usr/lib/hyprpolkitagent # Experimental Hyprland Polkit Agent
-exec-once = eww daemon && eww open bar # Daemon + Status Bar
-exec-once = dunst # Notification Daemon
-exec-once = hyprpaper # Wallpaper
-exec-once = hypridle # Idle daemon
-
-# Load pywal from cache
-exec-once = wal -R
-
-# Clipboard manager
-exec-once = wl-paste --type text --watch cliphist store # Stores text
-exec-once = wl-paste --type image --watch cliphist store # Stores images
-
-# Apps 
-exec-once = /bin/vesktop --start-minimized
-exec-once = /bin/steam-runtime -silent
