--- hypr/bindings.conf
@@ -1,117 +1,0 @@
-##############
-## BINDINGS ##
-##############
-
-# https://wiki.hyprland.org/Configuring/Keywords and https://wiki.hyprland.org/Configuring/Binds for information on how to configure input
-
-$terminal = kitty
-$fileManager = nautilus
-$menu = anyrun 
-$dmenu = anyrun --plugins libstdin.so
-$mainMod = SUPER
-$lockscreen = hyprlock
-$screenshotDir = $HOME/Screenshots
-$screenshotFull = hyprshot -m output -o $screenshotDir
-$screenshotSelect = hyprshot -m region -o $screenshotDir
-$media = spotify-launcher
-
-
-# Main binds, see https://wiki.hyprland.org/Configuring/Binds/ for more
-bind = $mainMod, K, exec, $terminal
-bind = $mainMod, Q, killactive
-bind = $mainMod, E, exec, $fileManager
-bind = $mainMod, F, togglefloating
-bind = $mainMod, SPACE, exec, $menu
-bind = $mainMod, P, pseudo, 
-bind = $mainMod, J, togglesplit
-bind = $mainMod, F11, fullscreen
-bind = $mainMod, N, exec, eww open --toggle control-center
-bind = $mainMod, L, exec, $lockscreen
-
-# Media keys
-bind = , XF86AudioMedia, exec, $media
-bind = , XF86AudioLowerVolume, exec, wpctl set-volume @DEFAULT_AUDIO_SINK@ 5%- # Decrease volume
-bind = , XF86AudioRaiseVolume, exec, wpctl set-volume @DEFAULT_AUDIO_SINK@ 5%+ # Increase volume
-bind = , XF86AudioMute, exec, wpctl set-mute @DEFAULT_AUDIO_SINK@ toggle # Mute
-bind = , XF86AudioPrev, exec, playerctl previous # Previous media
-bind = , XF86AudioNext, exec, playerctl next # Next media
-bind = , XF86AudioPlay, exec, playerctl play-pause # Toggle Play/Pause media
-
-# Brightness Keys
-bind = , XF86MonBrightnessDown, exec, brightnessctl s 5%- # Lower monitor brightness
-bind = , XF86MonBrightnessUp, exec, brightnessctl s +5% # Increase monitor brightness
-
-# Screenshot
-bind = , Print, exec, $screenshotSelect
-bind = $mainMod, Print, exec, $screenshotFull
-
-# Open clipboard
-bind = $mainMod, V, exec, cliphist list | $dmenu | cliphist decode | xargs -r wl-copy
-
-# Open wallpaper menu
-bind = $mainMod, W, exec, env bash $HOME/.config/hypr/scripts/change-wallpaper.sh
-
-# Reload binds
-# Eww
-bind = $mainMod, F7, exec, eww reload
-# Hyprpaper (Wallpaper)
-bind = $mainMod, F8, exec, pkill hyprpaper && hyprpaper
-# Dunst (Notification daemon)
-bind = $mainMod, F9, exec, pkill dunst && dunst
-
-# Move focus with mainMod + arrow keys
-bind = $mainMod, left, movefocus, l
-bind = $mainMod, right, movefocus, r
-bind = $mainMod, up, movefocus, u
-bind = $mainMod, down, movefocus, d
-
-# Move windows with keyboard keys
-bind = $mainMod SHIFT, left, movewindow, l
-bind = $mainMod SHIFT, right, movewindow, r
-bind = $mainMod SHIFT, up, movewindow, u
-bind = $mainMod SHIFT, down, movewindow, d 
-bind = $mainMod SHIFT, C, centerwindow
-
-# Resize windows with arrowkeys
-bind = $mainMod ALT, left, resizeactive, -60 0
-bind = $mainMod ALT, right, resizeactive, 60 0
-bind = $mainMod ALT, up, resizeactive, 0 -60
-bind = $mainMod ALT, down, resizeactive, 0 60
-
-# Switch workspaces with mainMod + [0-9]
-bind = $mainMod, 1, workspace, 1
-bind = $mainMod, 2, workspace, 2
-bind = $mainMod, 3, workspace, 3
-bind = $mainMod, 4, workspace, 4
-bind = $mainMod, 5, workspace, 5
-bind = $mainMod, 6, workspace, 6
-bind = $mainMod, 7, workspace, 7
-bind = $mainMod, 8, workspace, 8
-bind = $mainMod, 9, workspace, 9
-bind = $mainMod, 0, workspace, 10
-
-# Move active window to a workspace with mainMod + SHIFT + [0-9]
-bind = $mainMod SHIFT, 1, movetoworkspace, 1
-bind = $mainMod SHIFT, 2, movetoworkspace, 2
-bind = $mainMod SHIFT, 3, movetoworkspace, 3
-bind = $mainMod SHIFT, 4, movetoworkspace, 4
-bind = $mainMod SHIFT, 5, movetoworkspace, 5
-bind = $mainMod SHIFT, 6, movetoworkspace, 6
-bind = $mainMod SHIFT, 7, movetoworkspace, 7
-bind = $mainMod SHIFT, 8, movetoworkspace, 8
-bind = $mainMod SHIFT, 9, movetoworkspace, 9
-bind = $mainMod SHIFT, 0, movetoworkspace, 10
-
-bind = CTRL $mainMod, right, workspace, e+1
-bind = CTRL $mainMod, left, workspace, e-1
-
-bind = $mainMod, S, togglespecialworkspace, magic
-bind = $mainMod SHIFT, S, movetoworkspace, special:magic
-
-# Scroll through existing workspaces with mainMod + scroll
-bind = $mainMod, mouse_down, workspace, e+1
-bind = $mainMod, mouse_up, workspace, e-1
-
-# Move/resize windows with mainMod + LMB/RMB and dragging
-bindm = $mainMod, mouse:272, movewindow
-bindm = $mainMod, mouse:273, resizewindow
