--- hypr/decorations.conf
@@ -1,70 +1,0 @@
-
-################
-## DECORATION ##
-################
-
-# See https://wiki.hyprland.org/Configuring/Variables for more information on how to configure appearance
-
-source = ~/.cache/wal/colors-hypr.conf
-
-general {
-    gaps_in = 6
-    gaps_out = 8
-
-    border_size = 2
-
-    col.active_border = $color1
-    col.inactive_border = $background
-
-    resize_on_border = false
-
-    allow_tearing = false
-
-    layout = dwindle
-}
-
-misc {
-    animate_manual_resizes = false
-}
-
-decoration {
-    rounding = 14
-
-    # Active Window Opacity
-    active_opacity = 1.0
-    # Inactive Window Opacity
-    inactive_opacity = 0.95
-
-    shadow {
-        enabled = true
-        range = 3
-        render_power = 5
-        color = $background
-    }
-
-    blur {
-        enabled = true 
-        new_optimizations = true
-    	size = 7
-    	passes = 2
-    	vibrancy = 0.9
-    }
-}
-
-animations {
-    enabled = true
-
-    bezier = myBezier, 0.05, 0.9, 0.1, 1.05
-    bezier = amazingBezier, 0.25, 0.59, 0.1, 1.05
-    bezier = layerBezier, 0.36, 0.58, 0.1, 1
-
-    animation = windows, 1, 5, amazingBezier, slide
-    animation = windowsOut, 1, 5, amazingBezier, slide
-    animation = layers, 1, 6, layerBezier, slide
-    animation = layersOut, 1, 6, layerBezier, slide
-    animation = border, 1, 10, default
-    animation = borderangle, 1, 8, default
-    animation = fade, 1, 3.5, default
-    animation = fadeLayers, 1, 7, amazingBezier
-    animation = workspaces, 1, 4, amazingBezier, slidefade 20% 
-}
