--- hypr/environment.conf
@@ -1,17 +1,0 @@
-
-#################
-## ENVIRONMENT ##
-#################
-
-# Others
-env = XCURSOR_THEME, Adwaita
-env = XCURSOR_SIZE, 24
-env = HYPRCURSOR_THEME, Adwaita
-env = HYPRCURSOR_SIZE, 24
-env = QT_QPA_PLATFORM, wayland
-env = QT_QPA_PLATFORMTHEME, qt5ct
-env = QT_AUTO_SCREEN_SCALE_FACTOR, 1
-env = XDG_CURRENT_DESKTOP, Hyprland
-env = XDG_SESSION_TYPE, wayland
-env = ADW_DISABLE_PORTAL, 1
-env = MOZ_ENABLE_WAYLAND, 1
