--- hypr/hypridle.conf
@@ -1,13 +1,0 @@
-
-general {
-    lock_cmd = echo "Locked Hyprland Session"
-    unlock_cmd = echo "Unlocked Hyprland Session"
-    ignore_dbus_inhibit = false
-    ignore_systemd_inhibit = false
-}
-
-listener {
-    timeout = 3600 # 1800 -> 30m | 3600 -> 1h | 7200 -> 2h
-    on-timeout = hyprlock
-    on-resume = notify-send "Welcome back to Hyprland, $USER!"
-}
