--- hypr/hyprland.conf
@@ -1,37 +1,0 @@
-
-#############################################
-## Retrozinndev's Hyprland Configurations! ##
-#############################################
-
-# Nvidia Settings
-source = ~/.config/hypr/nvidia.conf
-
-# Environment
-source = ~/.config/hypr/environment.conf
-
-# Monitors
-source = ~/.config/hypr/monitors.conf
-
-# Layout
-source = ~/.config/hypr/layout.conf
-
-# Input
-source = ~/.config/hypr/input.conf
-
-# Devices
-# source = ~/.config/hypr/devices.conf # Uncomment this line to apply file, remember to make it first!
-
-# Plugins (you can comment if you want pure Hyprland)
-source = ~/.config/hypr/plugins.conf
-
-# Appearance
-source = ~/.config/hypr/decorations.conf
-
-# Autostart
-source = ~/.config/hypr/autostart.conf
-
-# Bindings
-source = ~/.config/hypr/bindings.conf
-
-# Rules
-source = ~/.config/hypr/rules.conf
