--- hypr/hyprlock.conf
@@ -1,95 +1,0 @@
-
-##############
-# LOCKSCREEN #
-##############
-
-# Source colors from pywal
-source = ~/.cache/wal/colors-hypr.conf
-
-# Fonts
-$font = Cantarell Regular
-$clockFont = Cantarell Black
-$minimalFont = Noto Sans Mono
-
-general {
-    disable_loading_bar = true
-    hide_cursor = false
-}
-
-background {
-    monitor =
-    path = $wallpaper
-    blur_passes = 2
-    color = $background
-}
-
-# Time
-label {
-    monitor =
-    text = cmd[update:30000] echo -e "$(date +"%R")" # 24-hours
-    # text = cmd[update:30000] echo -e "$(date +"%I:%M %p")" # 12-hours (AM/PM)
-    color = $foreground
-    font_size = 120
-    font_family = $clockFont
-    position = 0, -60
-    halign = center
-    valign = top
-}
-
-# Date
-label {
-    monitor = 
-    text = cmd[update:43200000] echo -e "$(date +"%A, %d %B %Y")"
-    color = $foreground
-    font_size = 20
-    font_family = $font
-    position = 0, -250
-    halign = center
-    valign = top
-}
-
-# Logged user
-label {
-    monitor = 
-    font_size = 6
-    font_family = $minimalFont
-    color = $foreground
-    text = Currently logged in as $USER
-    halign = center
-    valign = bottom
-    position = 0, 5
-}
-
-# Avatar
-image {
-    monitor = 
-    path = ~/.face
-    size = 72
-    border_color = $color2
-    position = 0, 100
-    halign = center
-    valign = bottom
-}
-
-# Input (password)
-input-field {
-    monitor =
-    size = 180, 35
-    outline_thickness = 2
-    dots_size = .15
-    dots_spacing = .6
-    dots_center = true
-    outer_color = $background
-    inner_color = $color3
-    font_color = $foreground
-    fade_on_empty = false
-    placeholder_text = 
-    hide_input = false
-    check_color = $color4
-    fail_color = $color1
-    fail_text = <i>$FAIL <b>($ATTEMPTS)</b></i>
-    capslock_color = $color1
-    position = 0, 40
-    halign = center
-    valign = bottom
-}
