--- hypr/hyprpaper.conf
@@ -1,6 +1,0 @@
-
-$wallpaper = /home/joaov/wallpapers/Miku City Sky.png
-
-splash = true
-preload = $wallpaper
-wallpaper = , $wallpaper
