--- hypr/input.conf
@@ -1,26 +1,0 @@
-###########
-## INPUT ##
-###########
-
-
-input {
-    kb_layout = br
-    kb_variant = abnt2
-    kb_model = pc105
-#   kb_options =
-#   kb_rules =
-
-    numlock_by_default = true
-    follow_mouse = 1
-
-    sensitivity = 0 # -1.0 - 1.0, 0 means no modification.
-
-    touchpad {
-        natural_scroll = true
-    }
-}
-
-# https://wiki.hyprland.org/Configuring/Variables/#gestures
-gestures {
-    workspace_swipe = true
-}
