--- hypr/layout.conf
@@ -1,4 +1,0 @@
-dwindle {
-    pseudotile = true
-    preserve_split = true 
-}
