--- hypr/monitors.conf
@@ -1,17 +1,0 @@
-
-##############
-## MONITORS ##
-##############
-
-# Configure yout monitor(s) here! See https://wiki.hyprland.org/Configuring/Monitors for more information on how to do that!
-
-# Monitor Arguments
-# arg0 -> monitor name(you can get monitor names with `hyprctl monitors`);
-# arg1 -> resolution@hertz;
-# arg2 -> positioning from the top-left corner;
-# arg3 -> scaling;
-# arg4 -> variable refresh rate for games(optional);
-# arg5 -> 1: vrr enabled, 0: no vrr.
-
-monitor = HDMI-A-1, 1920x1080@75, 0x0, 1, vrr, 1
-
