--- hypr/nvidia.conf
@@ -1,15 +1,0 @@
-
-#####################
-## NVIDIA SETTINGS ##
-#####################
-
-env = LIBVA_DRIVER_NAME, nvidia
-env = GBM_BACKEND, nvidia-drm
-env = __GLX_VENDOR_LIBRARY_NAME, nvidia
-env = _VK_LAYER_NV_optimus, NVIDIA_only
-env = __NV_PRIME_RENDER_OFFLOAD, 1
-
-cursor {
-    # Set to true if you have issues
-    no_hardware_cursors = true
-}
