--- hypr/plugins.conf
@@ -1,14 +1,0 @@
-####################
-# HYPRLAND PLUGINS #
-####################
-
-# You can add your preferred plugins here. Get help on how to do so: https://wiki.hyprland.org/Plugins/Using-Plugins/
-
-plugin {
-    # Example plugin configuration
-    #hyprbars {
-    #    bar_height = 24
-    #    hyprbars-button = rgb(ff4040), 16, 󰖭, hyprctl dispatch killactive
-    #    hyprbars-button = rgb(eeee11), 16, , hyprctl dispatch fullscreen 1
-    #}
-}
