--- hypr/rules.conf
@@ -1,72 +1,0 @@
-
-############################
-##  WINDOW / LAYER RULES  ##
-############################
-
-# See https://wiki.hyprland.org/Configuring/Window-Rules/ and https://wiki.hyprland.org/Configuring/Workspace-Rules/ for information on how to configure this
-
-# Floating windows
-windowrulev2 = float, class:moe.launcher.*
-windowrulev2 = float, class:com.github.rafostar.Clapper
-windowrulev2 = float, class:xdg-desktop-portal*
-windowrulev2 = float, class:org.pulseaudio.pavucontrol
-windowrulev2 = float, class:blueberry.py
-windowrulev2 = float, class:org.gnome.Loupe
-windowrulev2 = float, class:mcpelauncher-webview
-windowrulev2 = float, class:org.gnome.Calculator
-windowrulev2 = float, class:io.mrarm.mcpelauncher-ui-qt
-windowrulev2 = float, class:Resources
-
-# Resize
-windowrulev2 = size 50% 50%, class:org.pulseaudio.pavucontrol
-windowrulev2 = size 50% 50%, class:blueberry.py
-windowrulev2 = size 70% 70%, class:io.mrarm.mcpelauncher-ui-qt
-
-# Moving
-windowrulev2 = move 49.27% 7.28%, class:org.pulseaudio.pavucontrol
-windowrulev2 = move 49.27% 7.28%, class:blueberry.py
-windowrulev2 = movetoworkspace e, class:org.pulseaudio.pavucontrol
-
-# Animations
-windowrulev2 = animation slide right, class:org.pulseaudio.pavucontrol
-windowrulev2 = animation slide right, class:blueberry.py
-layerrule = animation slide right, swaync-control-center
-layerrule = animation fade, selection
-layerrule = animation fade, logout_dialog
-layerrule = animation fade, waybar
-layerrule = animation fade, hyprpaper
-layerrule = animation slide right, swaync-notification-window
-layerrule = animation fade, hyprpicker
-layerrule = animation fade, eww-calendar
-layerrule = animation fade, eww-audio
-
-# Opacity
-windowrulev2 = opacity .95 .95, class:kitty
-windowrulev2 = opacity .88 .88, class:spotify
-windowrulev2 = opacity .88 .88, class:hyprpolkitagent
-
-# No blur
-windowrulev2 = noblur, class:steam.*
-
-# Blur
-windowrulev2 = noblur, class:^()$, title:^()$
-
-# Window Blur list
-blurls = logout_dialog
-blurls = kitty
-blurls = eww-powermenu
-
-# Layer Blur list
-layerrule = blur, waybar
-layerrule = ignorealpha .5, waybar
-layerrule = blur, eww-bar
-layerrule = ignorealpha .5, eww-bar
-layerrule = blur, eww-calendar
-layerrule = ignorealpha .5, eww-calendar
-layerrule = blur, eww-cc
-layerrule = ignorealpha .5, eww-cc
-layerrule = blur, eww-audio
-layerrule = ignorealpha .5, eww-audio
-
-# Suppress maximize event from windows
-windowrulev2 = suppressevent maximize, class:.*
